+++ ost/s1/burst_batch.py
@@ -29,21 +29,35 @@ def bursts_to_ards(
 
     # we update max_workers in case we have less gpt_max_workers
     # then cpus available
-    if max_workers == 1 and config_dict['gpt_max_workers'] < os.cpu_count():
-        max_workers = int(os.cpu_count() / config_dict['gpt_max_workers'])
+    if max_workers == 1 or len(burst_gdf) == 1:
+        config_dict['gpt_max_workers'] = os.cpu_count()
+    elif len(burst_gdf) < max_workers:
+        config_dict['gpt_max_workers'] = int(max_workers / len(burst_gdf))
+        max_workers = int(max_workers / len(burst_gdf))
 
     out_files = {'bs': [], 'ls': [], 'coh': [], 'pol': []}
     # now we run with godale, which works also with 1 worker
-    executor = Executor(executor=executor_type, max_workers=max_workers)
-    for task in executor.as_completed(
-            func=burst_to_ard,
-            iterable=proc_inventory.iterrows(),
-            fargs=[config_dict]
-    ):
-        out_bs, out_ls, out_coh, out_pol = task.result()
-        out_files['bs'].append(out_bs)
-        out_files['ls'].append(out_ls)
-        out_files['coh'].append(out_coh)
-        out_files['pol'].append(out_pol)
+    if max_workers == 1 or len(burst_gdf) == 1:
+        for burst in proc_inventory.iterrows():
+            out_bs, out_ls, out_coh, out_pol = burst_to_ard(
+                burst=burst,
+                config_dict=config_dict
+            )
+            out_files['bs'].append(out_bs)
+            out_files['ls'].append(out_ls)
+            out_files['coh'].append(out_coh)
+            out_files['pol'].append(out_pol)
+    else:
+        executor = Executor(executor=executor_type, max_workers=max_workers)
+        for task in executor.as_completed(
+                func=burst_to_ard,
+                iterable=proc_inventory.iterrows(),
+                fargs=[config_dict]
+        ):
+            out_bs, out_ls, out_coh, out_pol = task.result()
+            out_files['bs'].append(out_bs)
+            out_files['ls'].append(out_ls)
+            out_files['coh'].append(out_coh)
+            out_files['pol'].append(out_pol)
 
     return out_files